tray_daemon/requirements.txt
# CloudToLocalLLM Enhanced Tray Daemon Dependencies
# Updated for Python 3.8+ compatibility with latest stable versions

# System tray functionality
pystray>=0.19.5

# Image processing for icons
Pillow>=11.0.0

# Process monitoring
psutil>=6.1.0

# HTTP client for connection testing
requests>=2.32.3

# Async HTTP client for connection broker
aiohttp>=3.11.10

# GUI toolkit (optional, for settings app)
PyGObject>=3.50.0
